Move SatelliteInfo back into UbxPacketNavSat.hpp (oops)
